completed alerts, pagination, softdeletes and additional features

+++ app/Http/Controllers/WeaponController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Models\Player;
 use App\Models\Weapon;
 use Illuminate\Http\Request;
 
@@ -16,42 +17,42 @@ class WeaponController extends Controller
      */
     public function weaponController()
     {
-        $weapons = Weapon::all();
+        $weapons = Weapon::withTrashed()->paginate(5);
         return view('weaponsPanel', ['weapons' => $weapons]);
     }
     public function destroy(string $id)
     {
         $weapon = Weapon::findOrFail($id);
-        $weapon->delete();
-        return redirect()->route('weaponsPanel');
+        $player = Player::where('weapon_id', $weapon->id)->exists();
+        if ($player) {
+            return redirect()->route('weaponsPanel')->with('assigned', 'Weapon assigned to player!');
+        } else {
+            $weapon->delete();
+            return redirect()->route('weaponsPanel')->with('delete', 'Weapon Deleted!');
+        }
     }
 
-    /**
-     * Store a newly created weapon in the database.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
+    public function restore(string $id)
+    {
+        $weapon = Weapon::withTrashed()->find($id);
+        $weapon->restore();
+        return redirect()->back()->with('restore', 'Weapon Restored!');
+    }
+
     public function addWeapon(Request $request)
     {
-        // Validate the request data
         $validatedData = $request->validate([
             'brand' => 'required|max:64',
             'model' => 'required|max:64',
             'type' => 'required|max:32',
         ]);
-
-        // Create a new weapon instance
         $weapon = new Weapon();
         $weapon->brand = $validatedData['brand'];
         $weapon->model = $validatedData['model'];
         $weapon->type = $validatedData['type'];
         $weapon->save();
-
-        // Redirect back with success message
-        return redirect()->route('weaponsPanel');
+        return redirect()->route('weaponsPanel')->with('add', 'Weapon Added!');
     }
-
 
     public function edit(string $id)
     {
@@ -64,7 +65,7 @@ class WeaponController extends Controller
         $weapon = Weapon::findOrFail($id);
         $input = $request->all();
         $weapon->update($input);
-        return redirect()->route('weaponsPanel')->with('flash_message', 'Weapon Updated!');
+        return redirect()->route('weaponsPanel')->with('update', 'Weapon Updated!');
     }
     public function showWeapons()
     {
